Note a moi même ne pas installer LiveWire dans un projet déjà créé...

+++ app/Http/Controllers/AjaxController.php
@@ -12,7 +12,7 @@ class AjaxController extends Controller
     public function fetchTasks(Request $request) {
         $startDate = $request->start ?? now()->startOfMonth();
         $endDate = $request->end ?? now()->endOfMonth();
-        $tasks = Task::whereBetween('due_date', [$startDate, $endDate])->get();
+        $tasks = Task::whereBetween('due_date', [$startDate, $endDate])->where('user_id', auth()->id())->get();
 
         $events = $tasks->map(function($task) {
             return [
@@ -20,9 +20,9 @@ class AjaxController extends Controller
                 'name' => $task->name,
                 'start' => $task->due_date,
                 // 'url' => route('task.show', $task->id)
-                'category_id' => $task->category->id, // 'category' is a custom property, not a native one
-                'category' => $task->category->name, // 'category' is a custom property, not a native one
-                'color' => $task->category->color,
+                'category_id' => $task->category ? $task->category->id : null,
+                'category' => $task->category ? $task->category->name : null,
+                'color' => $task->category ? $task->category->color : null,
                 'content' => $task->content ?? '',
                 'completed' => $task->completed ?? false,
             ];
@@ -34,6 +34,7 @@ class AjaxController extends Controller
     public function createTask(TaskRequest $taskRequest) {
         $task = new Task();
         $task->name = $taskRequest->name;
+        $task->user_id = auth()->id();
         $task->due_date = $taskRequest->due_date;
         $task->category_id = $taskRequest->category_id;
         $task->content = $taskRequest->content;
@@ -47,8 +48,18 @@ class AjaxController extends Controller
     }
 
     public function updateTask(TaskRequest $taskRequest) {
+        // Test if the user loggedin is the owner of the task
         $task = Task::find($taskRequest->id);
+        if ($task->user_id !== auth()->id()) {
+            return response()->json([
+                'status' => 'error',
+                'message' => 'You are not allowed to update this task',
+            ], 403);
+        }
+
+
         $task->name = $taskRequest->name;
+        $task->user_id = auth()->id();
         $task->due_date = $taskRequest->due_date ?? $task->due_date;
         $task->category_id = $taskRequest->category_id;
         $task->content = $taskRequest->content;
@@ -63,6 +74,12 @@ class AjaxController extends Controller
 
     public function deleteTask(Request $taskRequest) {
         $task = Task::find($taskRequest->id);
+        if ($task->user_id !== auth()->id()) {
+            return response()->json([
+                'status' => 'error',
+                'message' => 'You are not allowed to update this task',
+            ], 403);
+        }
         $task->delete();
         return response()->json([
             'status' => 'success',
